Change 'alpha' type from i16 to f32 in FIR filters
Modified the 'alpha' argument type in the 'generate_shape' method for FIR filters from i16 to f32 to allow more precise adjustments. This change affects all filter shape types and their corresponding tests. This adjustment in type allows for more precise control of the parameter and opens up a wider range of potential filtering effects.

+++ src/filters/fir/mod.rs
@@ -24,7 +24,7 @@ pub struct Windowing {
 
 impl Windowing {
     /// Generate a new windowing object
-    pub fn new(window_shape: WindowShapes, fft_size: usize, alpha: i16) -> Windowing {
+    pub fn new(window_shape: WindowShapes, fft_size: usize, alpha: f32) -> Windowing {
         // generate window
         let window = shapes::generate_shape(window_shape, fft_size, alpha);
